requirements.txt
undetected-chromedriver>=3.5.5,<4.0
selenium>=4.15,<5.0
deathbycaptcha-official>=5.0
